[net/http/SSPI] Move token length determination lower down the stack.
Currently the HttpAuthHandlerNegotiate::Factory() is responsible for
determining the maximum token size on Windows. This is done prior to
constructing the HttpNegotiateAuthSystem.

Conceptually the maximum token length is a property of the Negotiate
auth "system" (really should be called "mechanism"). Someone
constructing this should not have to know the token size. That's the
kind of thing that this abstraction layer is supposed to enapsulate.

Moreover, the "NegotiateLibrary" (i.e. the platform interface) currently
needs to be understood at both above and below the
HttpNegotiateAuthSystem abstraction layer.

In addition, to being an abstraction violation, having the library be
known above the HttpNegotiateAuthSystem layer means that the
functionality of the library needs to be accessible outside the
HttpNegotiateAuthSystem. In order for us to cleanly cleave out
privileged token operations out of the network process, we need to make
sure such functionality is neatly tucked away on one side of a single
abstraction. This is issue 991265.

Furthermore, having library access scattered outside
HttpNegotiateAuthSystem makes it difficult to move time consuming SSPI
operations into a separate thread. This is issue 711186.

To address these issues this CL moves the token size determination into
HttpNegotiateAuthSystem. Things still aren't quite there yet, but it's
a step in the right direction.

This change results in a behavior change. Currently the availability of
the Negotiate authentication mechanism is determined when constructing
the first HttpAuthHandlerNegotiate. With this change, the decision moves
to GenerateAuthToken(). Fortunately for us Chromium is not supported on
any Windows platform which doesn't have the Negotiate authentication
package. Hence there should not be any observable behavior change.

Bug: 991265, 711186
Change-Id: I9a30db15c6332a1304bc29112a0c26105a35282f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1814621
Reviewed-by: Ryan Sleevi <rsleevi@chromium.org>
Commit-Queue: Asanka Herath <asanka@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#705807}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: e7477a489ba72c6ddbb2b6dcd4f1346357dee75b