Allow extensions to modify CORS preflight with Web Request API (1/2)
Supporting preflight is tricky because preflight requests are made in
the network service, and they don't have request IDs. We are adding
a new message, OnLoaderForCorsPreflightCreated, to
TrustedURLLoaderHeaderClient to address that.

We are using TrustedURLLoaderHeaderClient, which means this mechanism
works only when "extraHeaders" is specified. Unlike for usual requests,
WebRequestProxyingUrlLoaderFactory::InProgressRequest dispatches
 - onBeforeRequest when OnLoaderForCorsPreflightCreated is called,
 - onBeforeSendHeaders and onSendHeaders when OnBeforeSendHeaders
   is called, and
 - onHeadersReceived, onResponseStarted, and onCompleted when
   OnHeadersReceived is called.

This CL is a preliminary change. It adds
 - TrustedURLLoaderHeaderClient.OnLoaderForCorsPreflightCreated
   as described above,
 - "remote_endpoint" param to TrustedHeaderClient.OnHeadersReceived
   to get endpoint information in OnHeadersReceived, and
 - kURLLoaderOptionAsCorsPreflight to mark a request as a CORS
   preflight.

This CL doesn't change any behavior because we have not set
kURLLoadOptionAsCorsPreflight and kURLLoadOptionUseHeaderClient in
network::cors::PreflightController yet.

The subsequent change is: https://crrev.com/c/1837561

Bug: 1002884
Change-Id: Ib396e35ce636f6ed016df089d393dedec94db430
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1833395
Commit-Queue: Yutaka Hirano <yhirano@chromium.org>
Reviewed-by: David Benjamin <davidben@chromium.org>
Reviewed-by: Karan Bhatia <karandeepb@chromium.org>
Reviewed-by: Kinuko Yasuda <kinuko@chromium.org>
Reviewed-by: Takashi Toyoshima <toyoshim@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#706372}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 832c8907284db238cec43d1ba0a3e8bb818c5cc3