Added metrics to track cross scheme SameSite cookie setting.
Add metric to track SameSite cookies set which cross schemes
http-to-https and https-to-http, as seen as site-for-cookies from
response and which are Lax or Strict.

Added histrogram to track number of SameSite cookies set with Lax or
Strict in the response.

Change-Id: I14668cc9ddf334b7b575e0199e407e389133de78
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1867205
Commit-Queue: Steven Bingler <bingler@chromium.org>
Reviewed-by: Lily Chen <chlily@chromium.org>
Reviewed-by: Charlie Harrison <csharrison@chromium.org>
Reviewed-by: Maksim Orlovich <morlovich@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#707133}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 33e3bac4ca9d4f665d8646a15e2f32666b2551c0