Add NetworkIsolationKey to HostCache::Key.
Currently an empty NetworkIsolationKey is always passed in in
production, so this should not caching behavior yet.

HostCache::GetAsListValue() and RestoreFromListValue() both read/write
these fields now, so there will be a one-time loss of cached host
entries.

Also fix a pair of SerializeAndDeserialize tests that were looking at
the HostCache the data was collect from, instead of the one that it
was copied to.

Bug: 997049
Change-Id: I16ac1247008b42398c23cd69a915e036a7c0ad7a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1880170
Commit-Queue: Matt Menke <mmenke@chromium.org>
Reviewed-by: Eric Orth <ericorth@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#709663}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 7b4a32f59926c5d534171944ef32f4efc9f2b0d2