[fuchsia] Fix NetworkChangeNotifier handling of initial state.
The NetworkChangeNotifier base class creates and initializes a
NetworkChangeCalculator which synthesizes high-level network
change notification events. Because it is created before the
derived NetworkChangeNotifierFuchsia reads the initial network
state, events must be dispatched for that initial state, for
the NetworkChangeCalculator to be up-to-date.

Bug: b/142796913
Change-Id: I51a319d0e5925721c81d7bb4b3eb15521dd1ba91
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1880449
Commit-Queue: Haines Sy <hainesy@google.com>
Reviewed-by: Wez <wez@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#709770}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 2eb17bcb2d6becb872cbedbb9c8c1ded535a3de0