Limit the number of reads in the gzip fuzzer.
It already has two timeout mitigations - one on the size of output, one
on the size of the input, but they're not enough. This CL replaces them
with a single limit on the number of reads, which should hopefully mean
we don't have to invest any more time dealing with silly timeouts.

Bug: 1006309
Change-Id: I50c6c52b26c43826efd5071ce0b345cd9782ca84
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1912142
Reviewed-by: David Benjamin <davidben@chromium.org>
Commit-Queue: Matt Menke <mmenke@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#714694}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: a1849a116101fa19520cc0fb1ecc325a8ed445f2