Make NetLog officially a global leaked object.
This allows NetLog to safely be used from worker threads that are not
joined on shutdown.

It already was a leaked object in Chromium and in Cronet on Android (but
not on Cronet iOS).

Moves ownership of the global object into net_log.cc and adds
NetLog::Get() to access it, and makes ~NetLog destructor private so that
code can't accidentally create local NetLog objects.

Tests can create a local NetLog with the TestNetLog class. Tests that
only needed an NetLog incidentally but don't actually observe/test it
are changed to use the global NetLog::Get().

Bug: 177538
Change-Id: I12a832801c108483db5b1a577c1094e7d2ceb3f0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1912696
Reviewed-by: Peter Beverloo <peter@chromium.org>
Reviewed-by: Wez <wez@chromium.org>
Reviewed-by: Clark DuVall <cduvall@chromium.org>
Reviewed-by: Jamie Walch <jamiewalch@chromium.org>
Reviewed-by: Rohit Rao <rohitrao@chromium.org>
Reviewed-by: Eric Roman <eroman@chromium.org>
Commit-Queue: Matt Mueller <mattm@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#719717}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: de5dadf50cde14c938f1a5cb4604960f7d82f5b7