Give site_for_cookies its own type, net/ portion
Previously it's been represented as a URL, which was misleading as it
does not have this sort of precision guaranteed, and is often empty.

Further, its computation and usage has special cases around non-http[s]
which makes a central spot to put them in valuable.

Non-coincidentally this change fixes SameSite=Lax cookies getting dropped
in Android Webview apps trying to set them on file:/// iframes.

This converts most[1] interfaces in net/, further CLs will push it further.
(Which also means this doesn't help that much, and there are some temporary
 conversions at edges, but it keeps the CL size sane).

[1] Exclusions are RedirectInfo, which makes sense to change the same time
as Network::ResourceRequest, as there is a lot of copying over, and
StaticCookiePolicy.

Bug: 577565, 1025349

Change-Id: I0886c53864f43f08fd7ed7a20a2dbc47d1397f95
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1925189
Reviewed-by: Andrey Kosyakov <caseq@chromium.org>
Reviewed-by: David Benjamin <davidben@chromium.org>
Reviewed-by: Eugene But <eugenebut@chromium.org>
Reviewed-by: Richard Coles <torne@chromium.org>
Reviewed-by: Lily Chen <chlily@chromium.org>
Commit-Queue: Maksim Orlovich <morlovich@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#723007}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 8be0e254edf7177fcbf3a46471cd11a21f2feb65