Make URLRequest::status() protected.
There are two problems with this method:

* It returns a URLRequestStatus.  This object has a network error code
and a 4-state value that summarizes the error code, and can be deduced
directly from the error code.  Many consumers have found this confusing,
and used URLRequestStatus enum values when they should have used
net::Errors, and vice versa.  URLRequestStatus::SUCCESS and net::OK are
both 0, making this even more likely to happen.

* It allows observing a URLRequest's final result before a request has
informed its caller of the result.  While this is much less concerning
than it once was, in the pre-network-service world, where a larger
number of consumeres dangled off ear URLRequest, it was possible for
those consumers to check this value, and react to it, assuming the
request was already complete, before the URLRequest::Delegate had
actually been informed of the failure, which could lead to bugs.

Ideally we'll eventually remove URLRequestStatus entirely, and replace
URLRequest::status_ with a single net::Error, but until then, this
CL largely addresses the issues with it, apart from URLFetcher
consumers, as that still provides status via a URLRequestStatus
object.

Bug: 651119
Change-Id: I6bf82a63088575e513013274e1e2e116a292831f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1958699
Reviewed-by: Lily Chen <chlily@chromium.org>
Reviewed-by: Eric Roman <eroman@chromium.org>
Reviewed-by: Tarun Bansal <tbansal@chromium.org>
Commit-Queue: Matt Menke <mmenke@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#725377}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 68813f0e5b5f1cf288fb7c474ff0e39964642669