A second attempt at fixing SimpleCacheCalculateSizeOfEntriesBetween
This is a second attempt to fix a unittest race.

The main thread gets a callback to update its total cache size that is
posted after the cache entry is created. Wait for that to be run
before checking the cache size.

I was able to reproduce the issue this time and can confirm that this
fixes it.

Bug: 1042441
Change-Id: Ic3f0f5a05415ba391be2887a3597c9645faac661
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2015738
Reviewed-by: Maksim Orlovich <morlovich@chromium.org>
Commit-Queue: Josh Karlin <jkarlin@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#734974}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: f8eb126a270c2bf830a64ad57b4bb61bb70a6047