Rename ProxyResolutionService to ConfiguredProxyResolutionService
This change is the first step in creating a proxy resolver that
completely relies on the underlying system to figure out a proxy. The
ProxyResolutionService is currently deeply tied to the notion of first
fetching proxy configurations (ex: a PAC URL) and then using those
configurations to resolve a proxy. Since a proxy resolver that truly
defers to the underlying operating system for all proxy logic does not
follow this model, it is helpful if we separate the generic
ProxyResolutionService from the notion of configurations.

This change renames the existing ProxyResolutionService, which relies
on fetching proxy configurations, to ConfiguredProxyResolutionService.
The only other changes to this file are suggestions from presubmit and
Tricium.

A set of future CLs will break up the ConfiguredProxyResolutionService
into the logical pieces needed to create a new, generic
ProxyResolutionService that does not necessarily depend on configs.

This change is intended to introduce zero functional changes. It is a
pure refactor change.

Bug: 1032820
Change-Id: I63c51c97e361d1cf5184e10ecb5f9f0deb1b5cdf
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2036408
Reviewed-by: Ramin Halavati <rhalavati@chromium.org>
Reviewed-by: Matt Menke <mmenke@chromium.org>
Reviewed-by: John Abd-El-Malek <jam@chromium.org>
Commit-Queue: Nicolas Arciniega <niarci@microsoft.com>
Cr-Original-Commit-Position: refs/heads/master@{#739597}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: d2013f9668d0587b1042572dadc6bd2e772d3248