Refactor ConfiguredProxyResolutionService::Request
This change moves out ConfiguredProxyResolutionService::Request into its
own file and renames it to ProxyResolutionRequest. This will be
necessary later on when we create a second version of the
ConfiguredProxyResolutionService under a base ProxyResolutionService
that relies completely on system APIs. In that second version, we will
also want to track the status of a proxy resolution request.

This change is intended to introduce zero functional changes. It is a
pure refactor change.

Bug: 1032820
Change-Id: I1f6192dc13289bb703ad2ea5ed3065e67c928fcd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2045147
Commit-Queue: Nicolas Arciniega <niarci@microsoft.com>
Reviewed-by: Matt Menke <mmenke@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#742379}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 9d383314fa8cbde76f36e5f9b138cd8a844f956c