Allow "*" as the leading character in structured header tokens.
This affects the parsing and serialization of tokens in structured
headers draft 15 and above. When operating in Draft 9 mode for backwards
compatibility, a character sequence beginning with an asterisk will be
parsed as a byte sequence instead.

Bug: 1048756
Change-Id: I5b1e2a57501a2e2dedb1183f495f71d75146c46b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2084711
Reviewed-by: Asanka Herath <asanka@chromium.org>
Commit-Queue: Ian Clelland <iclelland@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#746380}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 627af5f0648bb8bbf05da35871d801af8564ff62